requirements.txt
anyio==3.6.2
certifi==2022.6.15
click==8.1.3
distlib==0.3.6
fastapi==0.88.0
filelock==3.8.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pipenv==2022.9.4
platformdirs==2.5.2
pydantic==1.10.4
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
six @ file:///AppleInternal/BuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-110/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
watchfiles==0.18.1
websockets==10.4
